PoC: cli/new: Use bubbletea for template list
Switches the survey-based list picker we use to pick a template
to use bubbletea with the bubbles/list widget.
Instead of using any of the default bubbles/list renderings,
this implements a custom renderer for list items (the delegate)
that provides an experience closer in look to the 'pulumi new' today.

NOTE: This is just a PoC and not intended to be landed as is.
It hasn't been tested at all--minus some manual testing.
If y'all like this, in the future, some of the underlying components
can be generalized and re-used to replace other uses of survey.Select.

Related to #13815 in spirit.

A mergeable version of this would resolve #11812.