Change nodejs waiting on pipes to not panic
Hit this while testing with matrix testing. Fairly frequently the go
routine waiting for `pipesDone` would try to write to the response
channel and panic because it was already closed.

This rewrites the method to only use the response channel for the
running program, and we just explictly wait on all three events (pipes,
handle, program) and return the first one that triggers.

Also it refactors execNodejs to just be a sync method that returns the
response. The management of that running in a goroutine and writing to a
channel is just kept in Run.